# Conflicts:
#	dogfooding/lib/main.dart
#	dogfooding/lib/repos/app_repository.dart
#	dogfooding/lib/repos/auth_repo.dart
#	dogfooding/lib/repos/user_repository.dart
#	dogfooding/lib/src/screens/home_screen.dart

+++ packages/stream_video/lib/src/coordinator/open_api/coordinator_client_open_api.dart
@@ -23,18 +23,21 @@ import '../../models/queried_members.dart';
 import '../../models/user_info.dart';
 import '../../retry/retry_policy.dart';
 import '../../shared_emitter.dart';
+import '../../state_emitter.dart';
 import '../../token/token.dart';
 import '../../token/token_manager.dart';
 import '../../utils/none.dart';
 import '../../utils/result.dart';
 import '../../utils/standard.dart';
 import '../coordinator_client.dart';
+import '../models/coordinator_connection_state.dart';
 import '../models/coordinator_events.dart';
 import '../models/coordinator_models.dart';
 import 'coordinator_ws_open_api.dart';
 import 'open_api_extensions.dart';
 
 const _idEvents = 1;
+const _waitForConnectionTimeout = 5000;
 
 /// An accessor that allows us to communicate with the API around video calls.
 class CoordinatorClientOpenApi extends CoordinatorClient {
@@ -64,7 +67,13 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     basePath: _rpcUrl,
     authentication: _Authentication(
       apiKey: _apiKey,
-      tokenManager: _tokenManager,
+      getToken: () async {
+        final tokenResult = await _tokenManager.getToken();
+        if (tokenResult is! Success<UserToken>) {
+          throw (tokenResult as Failure).error;
+        }
+        return tokenResult.data;
+      },
       getConnectionId: () => _ws?.connectionId,
     ),
   );
@@ -76,24 +85,76 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   SharedEmitter<CoordinatorEvent> get events => _events;
   final _events = MutableSharedEmitterImpl<CoordinatorEvent>();
 
+  final _connectionState = MutableStateEmitterImpl<CoordinatorConnectionState>(
+    const CoordinatorDisconnected(),
+  );
+
   UserInfo? _user;
   CoordinatorWebSocketOpenApi? _ws;
   StreamSubscription<CoordinatorEvent>? _wsSubscription;
 
   @override
   Future<Result<None>> connectUser(UserInfo user) async {
-    _logger.d(() => '[connectUser] user: $user');
-    if (_user != null) {
-      _logger.w(() => '[connectUser] rejected (another user in use): $_user');
-      return Result.error(
-        'Another user is in use, please call "disconnectUser" first',
-      );
+    _logger.d(() => '[connectUser] user.id: ${user.id}');
+    final state = _connectionState.value;
+    if (state.isConnected) {
+      _logger.w(() => '[connectUser] rejected (already connected): $_user');
+      return const Result.success(none);
     }
+    if (state.isConnecting) {
+      _logger.w(() => '[connectUser] wait (already connecting): $_user');
+      return _waitUntilConnected();
+    }
+    _connectionState.value = CoordinatorConnectionState.connecting(
+      userId: user.id,
+    );
     _user = user;
     _ws = _createWebSocket(user).also((ws) {
-      _wsSubscription = ws.events.listen(_events.emit);
+      _wsSubscription = ws.events.listen((event) {
+        if (event is CoordinatorConnectedEvent) {
+          _logger.i(() => '[connectUser] WS connected');
+          _connectionState.value = CoordinatorConnectionState.connected(
+            userId: event.userId,
+            connectionId: event.connectionId,
+          );
+        } else if (event is CoordinatorDisconnectedEvent) {
+          _logger.i(() => '[connectUser] WS disconnected');
+          _connectionState.value = CoordinatorConnectionState.disconnected(
+            userId: event.userId,
+            connectionId: event.connectionId,
+            closeCode: event.closeCode,
+            closeReason: event.closeReason,
+          );
+        }
+        _events.emit(event);
+      });
     });
-    return openConnection();
+    final openResult = await openConnection();
+    if (openResult is Failure) {
+      _logger.e(() => '[connectUser] open failed: $openResult');
+      return openResult;
+    }
+    return _waitUntilConnected().whenComplete(() {
+      _logger.v(() => '[connectUser] completed');
+    });
+  }
+
+  Future<Result<None>> _waitUntilConnected() async {
+    _logger.w(() => '[waitUntilConnected] user.id: ${_user?.id}');
+    return _connectionState
+        .firstWhere(
+      (it) => it.isConnected,
+      // TODO
+      // replace timeout with config value,
+      timeLimit: const Duration(milliseconds: _waitForConnectionTimeout),
+    )
+        .then((it) {
+      _logger.v(() => '[waitUntilConnected] completed: $it');
+      return const Result.success(none);
+    }).onError((error, stackTrace) {
+      _logger.e(() => '[waitUntilConnected] failed: $error; $stackTrace');
+      return Result<None>.failure(VideoErrors.compose(error, stackTrace));
+    });
   }
 
   @override
@@ -179,6 +240,12 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     bool? voipToken,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[createDevice] no connection established');
+        return connectionResult;
+      }
+
       final input = open.CreateDeviceRequest(
         id: id,
         pushProvider: pushProvider,
@@ -208,6 +275,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   }) async {
     try {
       _logger.d(() => '[listDevices] userId: $userId');
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[listDevices] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.listDevices(
         userId: userId,
       );
@@ -230,6 +302,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   }) async {
     try {
       _logger.d(() => '[deleteDevice] id: $id, userId: $userId');
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[deleteDevice] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.deleteDevice(
         id: id,
         userId: userId,
@@ -258,6 +335,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
         () => '[getCall] cid: $callCid, ringing: $ringing'
             ', membersLimit: $membersLimit, ringing: $ringing, notify: $notify',
       );
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[getCall] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.getCall(
         callCid.type,
         callCid.id,
@@ -298,6 +380,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
         () => '[getOrCreateCall] cid: $callCid'
             ', ringing: $ringing, members: $members',
       );
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[getOrCreateCall] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.getOrCreateCall(
         callCid.type,
         callCid.id,
@@ -346,6 +433,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
         () => '[joinCall] cid: $callCid, dataCenterId: $datacenterId'
             ', ringing: $ringing, create: $create',
       );
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[joinCall] no connection established');
+        return connectionResult;
+      }
       final location = await _locationService.getLocation();
       _logger.v(() => '[joinCall] location: $location');
       final result = await _defaultApi.joinCall(
@@ -394,6 +486,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
         () => '[sendCustomEvent] cid: $callCid'
             ', eventType: $eventType, custom: $custom',
       );
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[sendCustomEvent] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.sendEvent(
         callCid.type,
         callCid.id,
@@ -424,6 +521,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
         () =>
             '[inviteUsers] cid: $callCid, members: $members, ringing: $ringing',
       );
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[inviteUsers] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.updateCallMembers(
         callCid.type,
         callCid.id,
@@ -447,6 +549,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     required List<CallPermission> permissions,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[requestPermissions] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.requestPermission(
         callCid.type,
         callCid.id,
@@ -471,6 +578,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     required List<CallPermission> revokePermissions,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[updateUserPermissions] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.updateUserPermissions(
         callCid.type,
         callCid.id,
@@ -492,6 +604,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<None>> startRecording(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[startRecording] no connection established');
+        return connectionResult;
+      }
       await _defaultApi.startRecording(callCid.type, callCid.id);
       return const Result.success(none);
     } catch (e, stk) {
@@ -505,6 +622,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     String sessionId,
   ) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[listRecordings] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.listRecordingsTypeIdSession1(
         callCid.type,
         callCid.id,
@@ -519,6 +641,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<None>> stopRecording(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[stopRecording] no connection established');
+        return connectionResult;
+      }
       await _defaultApi.stopRecording(callCid.type, callCid.id);
       return const Result.success(none);
     } catch (e, stk) {
@@ -529,6 +656,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<String?>> startBroadcasting(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[startBroadcasting] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi
           .startBroadcasting(callCid.type, callCid.id)
           .then((it) => it?.playlistUrl);
@@ -541,6 +673,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<None>> stopBroadcasting(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[stopBroadcasting] no connection established');
+        return connectionResult;
+      }
       await _defaultApi.stopBroadcasting(callCid.type, callCid.id);
       return const Result.success(none);
     } catch (e, stk) {
@@ -556,6 +693,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     Map<String, Object> custom = const {},
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[sendReaction] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.sendVideoReaction(
         callCid.type,
         callCid.id,
@@ -585,6 +727,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     int? limit,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[queryMembers] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.queryMembers(
         open.QueryMembersRequest(
           type: callCid.type,
@@ -615,6 +762,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     int? limit,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[queryCalls] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.queryCalls(
         open.QueryCallsRequest(
           filterConditions: filterConditions,
@@ -639,6 +791,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     required String userId,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[blockUser] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.blockUser(
         callCid.type,
         callCid.id,
@@ -659,6 +816,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     required String userId,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[unblockUser] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.unblockUser(
         callCid.type,
         callCid.id,
@@ -676,6 +838,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<None>> endCall(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[endCall] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.endCall(callCid.type, callCid.id);
       if (result == null) {
         return Result.error('endCall result is null');
@@ -694,6 +861,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     bool? startTranscription,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[goLive] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.goLive(
         callCid.type,
         callCid.id,
@@ -716,6 +888,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<CallMetadata>> stopLive(StreamCallCid callCid) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[stopLive] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.stopLive(callCid.type, callCid.id);
       if (result == null) {
         return Result.error('stopLive result is null');
@@ -737,6 +914,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     bool? screenshare,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[muteUsers] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.muteUsers(
         callCid.type,
         callCid.id,
@@ -772,6 +954,11 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     StreamGeofencingSettings? geofencing,
   }) async {
     try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[updateCall] no connection established');
+        return connectionResult;
+      }
       final result = await _defaultApi.updateCall(
         callCid.type,
         callCid.id,
@@ -829,7 +1016,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   }
 
   @override
-  Future<Result<GuestCreatedData>> createGuest({
+  Future<Result<GuestCreatedData>> loadGuest({
     required String id,
     String? name,
     String? role,
@@ -838,7 +1025,18 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     Map<String, Object> custom = const {},
   }) async {
     try {
-      final res = await _defaultApi.createGuest(
+      _logger.d(() => '[loadGuest] id: $id');
+      final defaultApi = open.DefaultApi(
+        open.ApiClient(
+          basePath: _rpcUrl,
+          authentication: _Authentication(
+            apiKey: _apiKey,
+            getToken: () => UserToken.anonymous(userId: id),
+            getConnectionId: () => _ws?.connectionId,
+          ),
+        ),
+      );
+      final result = await defaultApi.createGuest(
         open.CreateGuestRequest(
           user: open.UserRequest(
             id: id,
@@ -850,9 +1048,9 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
           ),
         ),
       );
-
-      if (res != null) {
-        return Result.success(res.toGuestCreatedData());
+      _logger.v(() => '[loadGuest] completed: $result');
+      if (result != null) {
+        return Result.success(result.toGuestCreatedData());
       } else {
         return const Result.failure(
           VideoError(message: 'Guest could not be created.'),
@@ -865,16 +1063,17 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
 }
 
 typedef GetConnectionId = String? Function();
+typedef GetToken = FutureOr<UserToken> Function();
 
 class _Authentication extends open.Authentication {
   _Authentication({
     required this.apiKey,
-    required this.tokenManager,
+    required this.getToken,
     required this.getConnectionId,
   });
 
   final String apiKey;
-  final TokenManager tokenManager;
+  final GetToken getToken;
   final GetConnectionId getConnectionId;
 
   @override
@@ -882,18 +1081,16 @@ class _Authentication extends open.Authentication {
     List<open.QueryParam> queryParams,
     Map<String, String> headerParams,
   ) async {
-    final tokenResult = await tokenManager.getToken();
-    if (tokenResult is! Success<UserToken>) {
-      throw (tokenResult as Failure).error;
-    }
     queryParams.add(open.QueryParam('api_key', apiKey));
     final connectionId = getConnectionId();
     if (connectionId != null) {
       queryParams.add(open.QueryParam('connection_id', connectionId));
     }
-    headerParams['Authorization'] = tokenResult.getDataOrNull()!.rawValue;
-    headerParams['stream-auth-type'] =
-        tokenResult.getDataOrNull()!.authType.name;
+    final userToken = await getToken();
+    headerParams['stream-auth-type'] = userToken.authType.name;
+    if (userToken.rawValue.isNotEmpty) {
+      headerParams['Authorization'] = userToken.rawValue;
+    }
     headerParams['X-Stream-Client'] = streamClientVersion;
     headerParams['x-client-request-id'] = const Uuid().v4();
   }
